Usability improvement

+++ app/src/main/java/com/termux/app/WearReceiverFragment.kt
@@ -1,9 +1,6 @@
 package com.termux.app
 
-import android.graphics.Bitmap
-import android.graphics.BitmapFactory
 import android.os.Bundle
-import android.os.Environment
 import android.os.SystemClock
 import android.view.InputDevice
 import android.view.KeyEvent
@@ -19,53 +16,73 @@ import com.google.android.gms.wearable.MessageClient
 import com.google.android.gms.wearable.MessageEvent
 import com.google.android.gms.wearable.Wearable
 import java.io.File
-import java.io.FileOutputStream
 import java.io.InputStream
 
 
-class WearReceiverFragment : Fragment(),
-    MessageClient.OnMessageReceivedListener,DataClient.OnDataChangedListener{
+class WearReceiverFragment : Fragment(), MessageClient.OnMessageReceivedListener,
+    DataClient.OnDataChangedListener {
     private var mActivity: TermuxActivity? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         mActivity = activity as TermuxActivity
-        Thread{
+        Thread {
             Tasks.await(Wearable.getNodeClient(mActivity!!).connectedNodes).forEach { it ->
                 Wearable.getMessageClient(mActivity!!)
                     .sendMessage(it.id, "/request-network", "".toByteArray()).addOnFailureListener {
-                    Toast.makeText(mActivity, "Failed to connect Mobile $it", Toast.LENGTH_SHORT).show()
-                    mActivity!!.supportFragmentManager.beginTransaction().remove(this).commitNow()
-                }
+                        Toast.makeText(
+                            mActivity, "Failed to connect Mobile $it", Toast.LENGTH_SHORT
+                        ).show()
+                        mActivity!!.supportFragmentManager.beginTransaction().remove(this)
+                            .commitNow()
+                    }
             }
         }
     }
 
     override fun onMessageReceived(p0: MessageEvent) {
-       Thread{ var text = String(p0.data)
-        when (p0.path) {
-            "/cmd" -> {
-                if (text.isEmpty())
-                    text = "\r"
-                mActivity!!.currentSession!!.write(text)
+        Thread {
+            var text = String(p0.data)
+            when (p0.path) {
+                "/cmd" -> {
+                    if (text.isEmpty()) text = "\r"
+                    mActivity!!.currentSession!!.write(text)
+                }
+
+                "/key" -> {
+                    val keys =
+                        text.splitToSequence(" ").filter { it.all { char -> char.isDigit() } }
+                            .toList()
+                    var keyMod = 0
+                    if (keys[1].toInt() == 1) keyMod = keyMod or KeyEvent.META_CTRL_ON
+                    if (keys[2].toInt() == 1) keyMod = keyMod or KeyEvent.META_ALT_ON
+                    if (keys[3].toInt() == 1) keyMod = keyMod or KeyEvent.META_SHIFT_ON
+                    val eventTime = SystemClock.uptimeMillis()
+                    mActivity!!.terminalView.dispatchKeyEvent(
+                        KeyEvent(
+                            eventTime,
+                            eventTime,
+                            KeyEvent.ACTION_DOWN,
+                            keys[0].toInt(),
+                            0,
+                            keyMod,
+                            0,
+                            0,
+                            KeyEvent.FLAG_SOFT_KEYBOARD or KeyEvent.FLAG_KEEP_TOUCH_MODE,
+                            InputDevice.SOURCE_KEYBOARD
+                        )
+                    )
+                }
             }
-            "/key" -> {
-                val keys =text.splitToSequence(" ").filter { it.all { char -> char.isDigit() } }.toList()
-                var keyMod = 0
-                if (keys[1].toInt()==1) keyMod = keyMod or KeyEvent.META_CTRL_ON
-                if (keys[2].toInt()==1) keyMod = keyMod or KeyEvent.META_ALT_ON
-                if (keys[3].toInt()==1) keyMod = keyMod or KeyEvent.META_SHIFT_ON
-                val eventTime= SystemClock.uptimeMillis()
-                mActivity!!.terminalView.dispatchKeyEvent(KeyEvent(eventTime,eventTime,KeyEvent.ACTION_DOWN,keys[0].toInt(),0,keyMod,0,0, KeyEvent.FLAG_SOFT_KEYBOARD or KeyEvent.FLAG_KEEP_TOUCH_MODE, InputDevice.SOURCE_KEYBOARD))
-            }
-        }
-       }.start()
+        }.start()
     }
 
     override fun onResume() {
         super.onResume()
         Wearable.getDataClient(mActivity!!).addListener(this)
-        Wearable.getMessageClient(requireContext()).addListener(this).addOnFailureListener { mActivity!!.supportFragmentManager.beginTransaction().remove(this).commitNow() }
+        Wearable.getMessageClient(requireContext()).addListener(this).addOnFailureListener {
+            mActivity!!.supportFragmentManager.beginTransaction().remove(this).commitNow()
+        }
     }
 
     override fun onPause() {
@@ -75,27 +92,36 @@ class WearReceiverFragment : Fragment(),
     }
 
     override fun onDataChanged(dataEvents: DataEventBuffer) {
-        dataEvents
-            .filter { it.type == DataEvent.TYPE_CHANGED && it.dataItem.uri.path == "/image" }
+        dataEvents.filter { it.type == DataEvent.TYPE_CHANGED }
             .forEach { event ->
-                val asset: Asset? = DataMapItem.fromDataItem(event.dataItem).dataMap.getAsset("profileImage")
-                Thread {
-                    val assetInputStream: InputStream? = Tasks.await(Wearable.getDataClient(mActivity!!).getFdForAsset(asset!!))?.inputStream
-                    val bitmap: Bitmap? =BitmapFactory.decodeStream(assetInputStream)
-                    mActivity!!.runOnUiThread { mActivity!!.showToast(if (bitmap==null)"bitmap is null" else "received",false) }
-                    val file = File(Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_PICTURES).absolutePath + "/wallpaper.jpeg")
+                if (event.dataItem.uri.path == "/files") {
+                    val fileAsset =
+                        DataMapItem.fromDataItem(event.dataItem).dataMap.getAsset("file")
+                    val filePath =
+                        DataMapItem.fromDataItem(event.dataItem).dataMap.getString("path")
+                    Thread {
+                        saveFileFromAsset(fileAsset!!, filePath!!)
+                    }.start()
+                    mActivity!!.showToast("↧", false)
+                } else if (event.dataItem.uri.path == "/delete") {
+                    val filePath =
+                        DataMapItem.fromDataItem(event.dataItem).dataMap.getString("path")
+                    val file = File(filePath!!)
                     if (file.exists()) file.delete()
-                    if (!file.parentFile?.exists()!!) file.mkdirs()
-                    file.createNewFile()
-                    val fos = FileOutputStream(file)
-                    bitmap!!.compress(Bitmap.CompressFormat.JPEG, 100, fos)
-                    fos.flush()
-                    fos.close()
-                    mActivity!!.runOnUiThread { mActivity!!.setWallpaper() }
-                    bitmap.recycle()
-                }.start()
+                    mActivity!!.showToast("deleted", false)
+                }
             }
         dataEvents.release()
     }
 
+    private fun saveFileFromAsset(asset: Asset, path: String) {
+        val assetInputStream: InputStream? =
+            Tasks.await(Wearable.getDataClient(mActivity!!).getFdForAsset(asset))?.inputStream
+        val file =
+            File(path)
+        if (file.exists()) file.delete()
+        if (!file.parentFile?.exists()!!) file.mkdirs()
+        file.createNewFile()
+        file.outputStream().use { assetInputStream!!.copyTo(it) }
+    }
 }
